+++ pkg/backend/httpstate/client/client.go
@@ -243,13 +243,27 @@ func (pc *Client) CreateStack(
 // DeleteStack deletes the indicated stack. If force is true, the stack is deleted even if it contains resources.
 func (pc *Client) DeleteStack(ctx context.Context, stack StackIdentifier, force bool) (bool, error) {
 	path := getStackPath(stack)
-	if force {
-		path += "?force=true"
+	queryObj := struct {
+		Force bool `url:"force"`
+	}{
+		Force: force,
 	}
 
-	// TODO[pulumi/pulumi-service#196] When the service returns a well known response for "this stack still has
-	//     resources and `force` was not true", we should sniff for that message and return a true for the boolean.
-	return false, pc.restCall(ctx, "DELETE", path, nil, nil, nil)
+	err := pc.restCall(ctx, "DELETE", path, queryObj, nil, nil)
+	return isStackHasResourcesError(err), err
+}
+
+func isStackHasResourcesError(err error) bool {
+	if err == nil {
+		return false
+	}
+
+	errRsp, ok := err.(*apitype.ErrorResponse)
+	if !ok {
+		return false
+	}
+
+	return errRsp.Code == 400 && errRsp.Message == "Bad Request: Stack still contains resources."
 }
 
 // EncryptValue encrypts a plaintext value in the context of the indicated stack.